v1.2 Rev 4
MASSIVE CHANGES!!!
This entire update is going to be huge!
Rev 4 adds the SmartInvs API, which allows us to easily draw Inventory GUIs for the user.
To play around with it a bit, I decided to make the entire configuration a GUI!
- Introducing /wwcc: modify your config.yml with an easy-to-use GUI.
- The conversations API is used in conjunction with SmartInvs to get user input. The next revision will hopefully have a GUI for /wwct and /wwcg!
- Loading invalid translator settings will not cause the plugin to disable anymore. Instead, it will load into a state where anything to do with translations is disabled. If the settings are valid, a simple /wwcr will put the plugin back into its normal state.
- Moved common code between /wwcr and onEnable() into one unified method, so that we don't have to keep modifying both.
- InventoryListener will listen for when our GUI is closed without saying; we will politely remind the user to save or reload the plugin.
- Update to /wwcs; use check and X marks instead of true/false
- prefixName is now WWC by default instead of "Default"...don't know why I did that in the first place tbh
And much more to come! v1.2 is going to take a while, but the payout should be a nice GUI with a bunch of squashed bugs!