Massively improve performance of cache
Finally! We now use the built-in methods that apparently average O(1) performance time (collisions prolly won't be an issue.)

This was done by changing the concurrenthashmap from CachedTranslation, Integer to CachedTranslation, Object[] and storing numberOfTimes/outputPhrase in the object array. This let us eliminate the for loop through the hashmap in CommonDefs :)

Also initialize hashmap with fixed size as defined in config. Prevents expanding the map's capacity completely and keeps that O(1) time complexity even more constant :)