Revamp cache algorithm
The last cache commit was awful. It solved nothing, added more issues, and was ultimately a 3AM commit that I deeply regret. The cache now uses a library built by a google engineer, and it does better than anything I could make. Everything works fine. There might be micro-optimizations, but we are finally in a good place.