Thorough rewrite...TODO: a lot of testing
Got switching between paper/spigot working!!!!!!!!! Took forever but worth it. Next commit will be expanding on this heavily, as well as hopefully adding some paper-exclusives (like hovering on outgoing messages and getting original translation, just like incoming)

A *lot* of testing is required, though. It is possible that GUI elements are broken, and the testing framework needs some expanding too. This will take a while!