Partial folia support
I will likely not spend much more time on this for a while, but here is a semi-functioning folia version of WWC. No GUIs work at all, many commands throw errors. The basic commands (such as /wwct en es without a GUI) as well as chat translation are functional. Reloading is broken, many tasks simply aren't scheduled, permissions for delays may not be properly handled, and there are likely many bugs that I have not adjusted for.