Begin adding separate chat channel option
This will benefit users who have a conflicting chat plugin.

My only major concern here is that I don't have time to implement proper filtering and become a full chat plugin. That is not the goal of this plugin. I will perhaps add the ability for ChatGPT to detect and block any weird messages, and tell the users to use 4o mini for the cheapest experience.